requirements.txt
absl-py==0.8.0
alembic==1.2.1
apache-airflow==1.10.5
apispec==3.0.0
astor==0.8.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
colorlog==4.0.2
croniter==0.3.30
cycler==0.10.0
dask==2.5.0
databricks-cli==0.9.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
docker==4.1.0
docutils==0.15.2
dumb-init==1.2.2
entrypoints==0.3
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Cors==3.0.8
Flask-JWT-Extended==3.23.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.2
fsspec==0.5.1
funcsigs==1.0.0
funcy==1.13
future==0.16.0
gast==0.3.2
gitdb2==2.0.6
GitPython==3.0.3
google-auth==1.6.3
google-pasta==0.1.7
gorilla==0.3.0
grpcio==1.24.1
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.0
json-merge-patch==0.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kubernetes==10.0.1
lazy-object-proxy==1.4.2
locket==0.2.0
lockfile==0.12.2
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.19.0
matplotlib==3.1.1
mistune==0.8.4
mlflow==1.3.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.15.4
oauthlib==3.1.0
onnx==1.5.0
ordereddict==1.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
pbr==5.4.3
pendulum==1.4.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prison==0.1.0
prometheus-client==0.7.1
prometheus-flask-exporter==0.9.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.3
ptyprocess==0.6.0
pyarrow==0.14.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pyrsistent==0.15.4
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.2
pytzdata==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
querystring-parser==1.2.4
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.12.0
sklearn==0.0
smmap2==2.0.5
SQLAlchemy==1.3.0
sqlparse==0.3.0
stevedore==1.31.0
tabulate==0.8.5
tb-nightly==1.14.0a20190603
tenacity==4.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
#tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tf-estimator-nightly==1.14.0.dev2019060501
thrift==0.11.0
toolz==0.10.0
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.6
utm==0.5.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
xmltodict==0.12.0
zope.deprecation==4.4.0
absl-py==0.8.0
alembic==1.2.1
apache-airflow==1.10.5
apispec==3.0.0
astor==0.8.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
colorlog==4.0.2
# configparser==4.0.2
configparser<3.6.0,>=3.5.0 
croniter==0.3.30
cycler==0.10.0
dask==2.5.0
databricks-cli==0.9.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
docker==4.1.0
docutils==0.15.2
dumb-init==1.2.2
entrypoints==0.3
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Cors==3.0.8
Flask-JWT-Extended==3.23.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.2
fsspec==0.5.1
funcsigs==1.0.0
funcy==1.13
future==0.16.0
gast==0.3.2
gitdb2==2.0.6
GitPython==3.0.3
google-auth==1.6.3
google-pasta==0.1.7
gorilla==0.3.0
grpcio==1.24.1
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.0
json-merge-patch==0.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kubernetes==10.0.1
lazy-object-proxy==1.4.2
locket==0.2.0
lockfile==0.12.2
lxml==4.4.1
Mako==1.1.0
markdown<3.0,>=2.5.2 
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.19.0
matplotlib==3.1.1
mistune==0.8.4
mlflow==1.3.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.15.4
oauthlib==3.1.0
onnx==1.5.0
ordereddict==1.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
pbr==5.4.3
pendulum==1.4.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prison==0.1.0
prometheus-client==0.7.1
prometheus-flask-exporter==0.9.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.3
ptyprocess==0.6.0
pyarrow==0.14.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
Pygments==2.4.2
PyJWT==1.7.1
#pyodbc==4.0.23
pyparsing==2.4.2
pyrsistent==0.15.4
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.2
pytzdata==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
querystring-parser==1.2.4
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.12.0
sklearn==0.0
smmap2==2.0.5
SQLAlchemy==1.3.0
sqlparse==0.3.0
stevedore==1.31.0
tabulate==0.8.5
tb-nightly==1.14.0a20190603
tenacity==4.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
#tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tf-estimator-nightly==1.14.0.dev2019060501
thrift==0.11.0
toolz==0.10.0
tornado<6.0.0,>=4.2.0 
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.6
utm==0.5.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
xmltodict==0.12.0
zope.deprecation==4.4.0
